requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-yasg==1.21.8
celery==5.4.0
gunicorn==23.0.0